032.redmine-mcp-server/requirements.txt
mcp>=0.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pydantic>=2.0.0
typing-extensions>=4.0.0
python-dotenv>=1.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0